# lib/api-zod/src/generated/types/battleFighter.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Creature Collection Game API
 * OpenAPI spec version: 0.1.0
 */
import type { BattleFighterRarity } from './battleFighterRarity';
import type { BattleFighterType } from './battleFighterType';

export interface BattleFighter {
  id: number;
  name: string;
  rarity: BattleFighterRarity;
  type: BattleFighterType;
  attack: number;
  defense: number;
  speed: number;
  health: number;
  powerScore: number;
  skillName: string;
  skillDescription: string;
}