# lib/api-zod/src/generated/types/battleLogEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Creature Collection Game API
 * OpenAPI spec version: 0.1.0
 */
import type { BattleLogEntryType } from './battleLogEntryType';

export interface BattleLogEntry {
  turn: number;
  message: string;
  type: BattleLogEntryType;
}