# lib/api-zod/src/generated/types/battleResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Creature Collection Game API
 * OpenAPI spec version: 0.1.0
 */
import type { BattleCreatureSummary } from './battleCreatureSummary';
import type { BattleLogEntry } from './battleLogEntry';
import type { BattleResultOutcome } from './battleResultOutcome';

export interface BattleResult {
  outcome: BattleResultOutcome;
  coinsEarned: number;
  logEntries: BattleLogEntry[];
  playerTeam: BattleCreatureSummary[];
  opponentTeam: BattleCreatureSummary[];
}