# lib/api-zod/src/generated/types/inventoryStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Creature Collection Game API
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryStatsByRarity } from './inventoryStatsByRarity';
import type { InventoryStatsStrongestCreature } from './inventoryStatsStrongestCreature';

export interface InventoryStats {
  total: number;
  byRarity: InventoryStatsByRarity;
  totalPower: number;
  /** @nullable */
  strongestCreature: InventoryStatsStrongestCreature;
}