# lib/api-zod/src/generated/types/listingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Creature Collection Game API
 * OpenAPI spec version: 0.1.0
 */

export interface ListingInput {
  ownedCreatureId: number;
  price: number;
}